Remove downcast-selection-converters.js.
